# [SSCP] ranged parallel_for: On CPU, check whether WGs are in the domain only if they are at the edges