As found out in #1852, there can be a noticable performance hit in `parallel_for(range)` compared to `parallel_for(nd_range)` because `parallel_for(range)` needs to also support cases where the `range` is not divisible by the work group size, in which case the last work groups will only partially participate in the computation.

This requires an additional if statement to check whether items are still part of the parallel domain. This if statement can negatively affect the generated code.

This PR changes the logic, such that on CPU, we only do this check for the last work groups in each dimension. This allows us to have most of the work groups without the check (and thus be more efficient). This approach is possible because only the last work groups in each dimension can potentially exceed the iteration space.

@blinkfrog With this change, perf improves from ~78ms to 58ms for your original benchmark code for me! Also, OpenCL on CPU improves from 225 to ~200ms.