@illuhad 2026-09-04
> Is something blocking this?

No, I just forgot pushing it :)

@sbalint98, can you review? This PR has potential performance impact for the SSCP->CPU path.